+++ ios/AppLovinMAX.m
@@ -7,6 +7,7 @@
 //
 
 #import "AppLovinMAX.h"
+#import "AppLovinMAXNativeAdView.h"
 
 #define ROOT_VIEW_CONTROLLER (UIApplication.sharedApplication.keyWindow.rootViewController)
 
@@ -731,6 +732,10 @@ RCT_EXPORT_METHOD(setRewardedAdExtraParameter:(NSString *)adUnitIdentifier :(NSS
     {
         name = @"OnRewardedAdLoadedEvent";
     }
+    else if ( MAAdFormat.native == adFormat )
+    {
+        name = @"OnNativeAdLoadedEvent";
+    }
     else
     {
         [self logInvalidAdFormat: adFormat];
@@ -763,8 +768,15 @@ RCT_EXPORT_METHOD(setRewardedAdExtraParameter:(NSString *)adUnitIdentifier :(NSS
     }
     else
     {
-        [self log: @"invalid adUnitId from %@", [NSThread callStackSymbols]];
-        return;
+        if ( [adUnitIdentifier isEqualToString: AppLovinMAXNativeAdView.errorAdUnitIdentifier] )
+        {
+            name = @"OnNativeAdLoadFailedEvent";
+        }
+        else
+        {
+            [self log: @"invalid adUnitId from %@", [NSThread callStackSymbols]];
+            return;
+        }
     }
     
     [self sendReactNativeEventWithName: name body: @{@"adUnitId" : adUnitIdentifier,
@@ -793,6 +805,10 @@ RCT_EXPORT_METHOD(setRewardedAdExtraParameter:(NSString *)adUnitIdentifier :(NSS
     else if ( MAAdFormat.rewarded == adFormat )
     {
         name = @"OnRewardedAdClickedEvent";
+    }
+    else if ( MAAdFormat.native == adFormat )
+    {
+        name = @"OnNativeAdClickedEvent";
     }
     else
     {
@@ -909,6 +925,10 @@ RCT_EXPORT_METHOD(setRewardedAdExtraParameter:(NSString *)adUnitIdentifier :(NSS
     else if ( MAAdFormat.rewarded == adFormat )
     {
         name = @"OnRewardedAdRevenuePaid";
+    }
+    else if ( MAAdFormat.native == adFormat )
+    {
+        name = @"OnNativeAdRevenuePaid";
     }
     else
     {
@@ -1574,7 +1594,12 @@ RCT_EXPORT_METHOD(setRewardedAdExtraParameter:(NSString *)adUnitIdentifier :(NSS
              @"OnRewardedAdFailedToDisplayEvent",
              @"OnRewardedAdHiddenEvent",
              @"OnRewardedAdReceivedRewardEvent",
-             @"OnRewardedAdRevenuePaid"];
+             @"OnRewardedAdRevenuePaid",
+             
+             @"OnNativeAdLoadedEvent",
+             @"OnNativeAdLoadFailedEvent",
+             @"OnNativeAdClickedEvent",
+             @"OnNativeAdRevenuePaid"];
 }
 
 - (void)startObserving